+++ ksef-client/src/main/java/pl/akmf/ksef/sdk/system/KsefIntegrationMode.java
@@ -1,0 +1,6 @@
+package pl.akmf.ksef.sdk.system;
+
+public enum KsefIntegrationMode {
+    OFFLINE,
+    ONLINE
+}